Refactor bigtable/testing/random.h to google/cloud.
This fixes #681. We need (pseudo) random number generators for the
same reason the Bigtable client does: to generate random data, and
random names for buckets and objects.
