+++ google/cloud/bigtable/benchmarks/bigtable_benchmark_test.cc
@@ -69,7 +69,7 @@ TEST(BenchmarkTest, MakeRandomKey) {
   BenchmarkSetup setup("key", argc, argv);
 
   Benchmark bm(setup);
-  auto gen = google::cloud::bigtable::testing::MakeDefaultPRNG();
+  auto gen = google::cloud::MakeDefaultPRNG();
 
   // First make sure that the keys are not always the same.
   auto make_some_keys = [&bm, &gen]() {
