+++ google/cloud/bigtable/bigtable_client_unit_tests.bzl
@@ -31,7 +31,6 @@ bigtable_client_unit_tests = [
     "table_sample_row_keys_test.cc",
     "table_test.cc",
     "table_readmodifywriterow_test.cc",
-    "testing/random_test.cc",
     "read_modify_write_rule_test.cc",
     "row_reader_test.cc",
     "row_test.cc",
