+++ google/cloud/bigtable/testing/table_integration_test.cc
@@ -128,8 +128,8 @@ void TableIntegrationTest::CheckEqualUnordered(
 std::string TableIntegrationTest::RandomTableId() {
   constexpr int RANDOM_CHARACTERS = 8;
   return std::string("table-") +
-         bigtable::testing::Sample(generator_, RANDOM_CHARACTERS,
-                                   "abcdefghijklmnopqrstuvwxyz0123456789");
+         google::cloud::Sample(generator_, RANDOM_CHARACTERS,
+                               "abcdefghijklmnopqrstuvwxyz0123456789");
 }
 }  // namespace testing
