+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -2,6 +2,7 @@
 google_cloud_cpp_common_unit_tests = [
     "internal/optional_test.cc",
     "internal/throw_delegate_test.cc",
+    "random_test.cc",
     "log_test.cc",
 ]
