+++ google/cloud/random.cc
@@ -12,12 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/bigtable/testing/random.h"
+#include "google/cloud/random.h"
 
 namespace google {
 namespace cloud {
-namespace bigtable {
-namespace testing {
+inline namespace GOOGLE_CLOUD_CPP_NS {
 std::string Sample(DefaultPRNG& gen, int n, std::string const& population) {
   std::uniform_int_distribution<std::size_t> rd(0, population.size() - 1);
 
@@ -27,7 +26,6 @@ std::string Sample(DefaultPRNG& gen, int n, std::string const& population) {
   return result;
 }
 
-}  // namespace testing
-}  // namespace bigtable
+}  // namespace GOOGLE_CLOUD_CPP_NS
 }  // namespace cloud
 }  // namespace google
